v1.7.7

+++ src/main/java/org/dimasik/liteauction/frontend/commands/impl/Sell.java
@@ -176,7 +176,7 @@ public class Sell extends SubCommand {
         switch (args.length){
             case 2:
                 try{
-                    int cnt = Integer.parseInt(lastArg);
+                    double cnt = Double.parseDouble(lastArg);
                     completions.add(cnt + "k");
                     completions.add(cnt + "kk");
                     completions.add(cnt + "m");
